Create service for cross cutting concerns

+++ book-microservice/src/main/java/com/edu/bookmicroservice/repo/BookRepository.java
@@ -1,16 +1,17 @@
 package com.edu.bookmicroservice.repo;
 
-import com.edu.bookmicroservice.model.Book;
+
+import com.edu.bookmicroservice.entity.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
+
 import java.util.Optional;
 
 @Repository
-public interface BookRepository extends JpaRepository<Book, Long> {
+public interface BookRepository extends JpaRepository<Book, Integer> {
 
     @Query("SELECT bk FROM Book bk WHERE bk.isbn=(:isbn)")
     Optional<Book> findByIsbn(@Param("isbn") String isbn);
